requirements.txt
numpy
opencv-python
pillow
torch
torchvision
pynput
